Follower works pretty well, jerk profiles are next
- dynamically change text size for Command name in UI bar based on
length (g.fontMetrics() trick)
- calculate trajectory output based on model?
  ie.
At this time, I am at this position, this velocity and this
acceleration. At this next time, I need to be at this position, this
velocity and this acceleration. Knowing my size, mass, etc., what
voltage do I need to output to get to that state?
- 5-segment jerk profile afterwards